backend/requirements.txt
# 华图儿AI创意绘画应用 - Python依赖包

# Web框架
Flask==2.3.3
Flask-CORS==4.0.0

# 异步任务处理
celery==5.3.4
redis==5.0.1

# 数据库和存储
supabase==2.0.2
psycopg2-binary==2.9.7

# HTTP请求
requests==2.31.0

# 图像处理
Pillow==10.0.1

# 工具库
python-dotenv==1.0.0
python-multipart==0.0.6

# 日志和监控
gunicorn==21.2.0

# 开发工具
pytest==7.4.3
pytest-flask==1.3.0
black==23.9.1
flake8==6.1.0
